refactor(domain): align entities and DTOs for user and requirement flow

+++ Backend/src/main/java/com/afk/control/dto/TipoRequisitoDto.java
@@ -1,6 +1,6 @@
 package com.afk.control.dto;
 
 public record TipoRequisitoDto(
-        Integer id,
+        long id,
         String nombre
 ) {}
